+++ app/Models/Tag.php
@@ -6,4 +6,5 @@ use Niang\Core\Database\Model;
 
 class Tag extends Model
 {
+    protected static array $fillable = ['name'];
 }